Create a "Workflow" for `CreateMov`
This introduces a slight structure to AE Submitter features, starting with the new `CreateMOV` feature.

We build a small module in `Workflows/<WorkflowName>/WorkflowName.jsx`.
A workflow module has two functions:
- `runChecks(renderQueueItem: renderQueueItem)`
- `generateRunScript(renderQueueItem: renderQueueItem) -> string`

The `runChecks` is a function that checks that a passed renderQueueItem is in a valid state to operate the workflow. For example, our `CreateMov` workflow requires the output of the RQI to be a frame sequence. If the workflow.runChecks determines that the RenderQueueItem does not meet the requirements, it should `alert` a message explaining so and then throw that message so that the submission does not continue.

`generateRunScript` will generate Powershell code that can be used in a runStep specifically for this workflow. This concept may need a little massaging if we expand the "workflow" concept to other features. We may need to either return an entire step (as opposed to just the embeddedFiles.data), and potentially take the current template in as an argument?